Full-Duplex_Interaction/baseline/requirements.txt
absl-py==2.2.2
accelerate==1.5.2
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiohttp-cors==0.8.0
aiosignal==1.3.2
airportsdata==20250224
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.2
anyio==4.9.0
astor==0.8.1
attrs==25.3.0
audioread==3.0.1
av==14.2.0
beautifulsoup4==4.13.3
bidict==0.23.1
bitsandbytes==0.45.3
blake3==1.0.4
blessed==1.20.0
blinker==1.9.0
cachetools==5.5.2
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chattts==0.2.3
click==8.1.8
cloudpickle==3.1.1
cn2an==0.5.23
codetiming==1.4.0
colorama==0.4.6
coloredlogs==15.0.1
colorful==0.5.6
compressed-tensors==0.9.2
conformer==0.3.2
contourpy==1.3.1
contractions==0.1.73
crcmod==1.7
cryptography==44.0.2
cssselect2==0.8.0
cupy-cuda12x==13.4.0
cut-cross-entropy==25.1.1
cycler==0.12.1
cytoolz==1.0.1
datasets==3.4.1
decorator==5.2.1
decord==0.6.0
deepspeed==0.16.5
defusedxml==0.7.1
depyf==0.18.0
diffusers==0.32.2
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docker-pycreds==0.4.0
docstring_parser==0.16
easydict==1.13
editdistance==0.8.1
einops==0.8.1
einx==0.3.0
ema-pytorch==0.7.7
email_validator==2.2.0
encodec==0.1.1
et_xmlfile==2.0.0
fastapi==0.115.11
fastapi-cli==0.0.7
fastrlock==0.8.3
ffmpeg==1.4
ffmpy==0.5.0
filelock==3.18.0
fire==0.7.0
Flask==3.1.0
Flask-SocketIO==5.5.1
flatbuffers==25.2.10
fonttools==4.56.0
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.12.0
funasr==1.2.6
gdown==5.2.0
gguf==0.10.0
gh==0.0.4
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
gpustat==1.1.1
gradio==5.22.0
gradio_client==1.8.0
graphviz==0.20.3
groovy==0.1.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
hf_transfer==0.1.9
hjson==3.1.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.3
humanfriendly==10.0
hydra-core==1.3.2
HyperPyYAML==1.2.2
idna==3.10
imagesize==1.4.1
importlib_metadata==8.6.1
inflect==7.5.0
interegular==0.3.3
intervaltree==3.1.0
itsdangerous==2.2.0
jaconv==0.4.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
jiter==0.9.0
jiwer==3.1.0
jmespath==0.10.0
joblib==1.4.2
jq==1.8.0
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaldialign==0.9.1
kaldiio==2.18.1
kiwisolver==1.4.8
lark==1.2.2
latex2sympy2_extended==1.10.1
lazy_loader==0.4
librosa==0.11.0
lightning==2.5.1
lightning-utilities==0.14.3
lilcom==1.8.1
llvmlite==0.43.0
lm-format-enforcer==0.10.11
loguru==0.7.3
lxml==5.3.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
math-verify==0.7.0
matplotlib==3.10.1
mdurl==0.1.2
mistral_common==1.5.4
modelscope==1.24.1
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.2.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
onnxruntime==1.21.0
openai==1.67.0
openai-whisper==20240930
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.10.15
oss2==2.19.1
outlines==0.1.11
outlines_core==0.1.26
packaging==24.2
pandas==2.2.3
partial-json-parser==0.2.1.1.post5
peft==0.15.0
pillow==11.1.0
platformdirs==4.3.7
pooch==1.8.2
portalocker==2.0.0
proces==0.1.7
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.21.1
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
py-spy==0.4.0
pyahocorasick==2.1.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pybase16384==0.3.8
pybind11==2.13.6
pycocoevalcap==1.2
pycocotools==2.0.8
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pylatexenc==2.10
pynndescent==0.5.13
pyparsing==3.2.1
pypinyin==0.54.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-json-logger==3.3.0
python-multipart==0.0.20
python-socketio==5.12.1
pytorch-lightning==2.5.1
pytorch-wpe==0.0.1
pytz==2025.1
pyworld==0.3.5
PyYAML==6.0.2
pyzmq==26.3.0
qa_metrics==0.2.31
quanto==0.2.0
RapidFuzz==3.12.2
ray==2.43.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rotary-embedding-torch==0.8.6
rpds-py==0.23.1
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
ruff==0.11.0
sacrebleu==1.5.1
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.15.2
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.24.1
setproctitle==1.3.5
setuptools==75.8.0
shellingham==1.5.4
shtab==1.7.1
silero-vad==5.1.2
simple-websocket==1.1.0
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
speechbrain==1.0.2
starlette==0.46.1
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensordict==0.6.2
termcolor==2.5.0
textsearch==0.0.24
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.19.1
tomli==2.2.1
tomlkit==0.13.2
toolz==1.0.0
torch==2.6.0
torch-complex==0.4.4
torchaudio==2.6.0
torchdata==0.11.0
torchdiffeq==0.2.5
torchmetrics==1.7.1
torchvision==0.21.0
tqdm==4.67.1
transformers==4.44.0
triton==3.2.0
typeguard==4.4.2
typer==0.15.2
typing_extensions==4.12.2
tyro==0.9.17
tzdata==2025.1
ujson==5.10.0
umap-learn==0.5.7
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
vector-quantize-pytorch==1.22.2
verl==0.3.0.post0
virtualenv==20.29.3
vocos==0.1.0
wandb==0.19.8
watchfiles==1.0.4
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
wget==3.2
wheel==0.45.1
whisper==1.1.10
wrapt==1.17.2
wsproto==1.2.0
x-transformers==2.2.12
xformers==0.0.29.post2
xgrammar==0.1.16
xxhash==3.5.0
yarl==1.18.3
zhconv==1.4.3
zipp==3.21.0
lhotse==1.7.0
websocket-client==1.9.0
gxl_ai_utils @ git+https://gitee.com/geng-xuelong/gxl_ai_utils.git@7c10a934a9c154c09dbc8785e32cf3be77598c30
